add product to user cart from SpecificProductFragment.kt

+++ app/src/main/java/com/mhmdawad/superest/data/repository/ShopRepository.kt
@@ -171,12 +171,13 @@ constructor(
 
     }
 
-    suspend fun addProductsToCart(list: List<ProductModel>): Resource<Any> {
+    suspend fun addProductsToCart(list: List<ProductModel>, deleteFavoriteProducts: Boolean): Resource<Any> {
         return try {
             list.forEach { product ->
                 cartCollection.document(product.id).set(product).await()
             }
-            favoriteDao.deleteAllProducts()
+            if(deleteFavoriteProducts)
+                favoriteDao.deleteAllProducts()
             Resource.Success(Any())
         } catch (e: Exception) {
             Resource.Error(errorMessage)
